llm相关对象、序列化代码放到core包
