+++ spring-ai-alibaba-graph/spring-ai-alibaba-graph-core/src/main/java/com/alibaba/cloud/ai/graph/serializer/agent/AgentAction.java
@@ -1,0 +1,7 @@
+package com.alibaba.cloud.ai.graph.serializer.agent;
+
+import org.springframework.ai.chat.messages.AssistantMessage;
+
+public record AgentAction(AssistantMessage.ToolCall toolCall, String log) {
+
+}
