+++ spring-ai-alibaba-graph/spring-ai-alibaba-graph-core/src/main/java/com/alibaba/cloud/ai/graph/serializer/agent/AgentActionDeserializer.java
@@ -1,10 +1,10 @@
-package dev.ai.alibaba.samples.executor.std.json;
+package com.alibaba.cloud.ai.graph.serializer.agent;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
-import dev.ai.alibaba.samples.executor.AgentAction;
+import org.springframework.ai.chat.messages.AssistantMessage;
 
 import java.io.IOException;
 
@@ -13,7 +13,14 @@ class AgentActionDeserializer extends JsonDeserializer<AgentAction> {
 	@Override
 	public AgentAction deserialize(JsonParser parser, DeserializationContext ctx) throws IOException {
 		JsonNode node = parser.getCodec().readTree(parser);
-		return new AgentAction(node.get("log").asText());
+
+		var toolExecutionRequestNode = node.get("toolExecutionRequest");
+		var toolExecutionRequest = ctx.readValue(toolExecutionRequestNode.traverse(parser.getCodec()), AssistantMessage.ToolCall.class);
+
+		return new AgentAction(
+				toolExecutionRequest,
+				node.get("log").asText()
+		);
 	}
 
 }
