+++ spring-ai-alibaba-graph/spring-ai-alibaba-graph-core/src/main/java/com/alibaba/cloud/ai/graph/serializer/agent/AgentFinishDeserializer.java
@@ -1,11 +1,10 @@
-package dev.ai.alibaba.samples.executor.std.json;
+package com.alibaba.cloud.ai.graph.serializer.agent;
 
 import com.fasterxml.jackson.core.JacksonException;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
-import dev.ai.alibaba.samples.executor.AgentFinish;
 
 import java.io.IOException;
 import java.util.HashMap;
