+++ spring-ai-alibaba-graph/spring-ai-alibaba-graph-core/src/main/java/com/alibaba/cloud/ai/graph/serializer/agent/AgentOutcome.java
@@ -1,4 +1,4 @@
-package dev.ai.alibaba.samples.executor;
+package com.alibaba.cloud.ai.graph.serializer.agent;
 
 public record AgentOutcome(AgentAction action, AgentFinish finish) {
 }
