+++ spring-ai-alibaba-graph/spring-ai-alibaba-graph-core/src/main/java/com/alibaba/cloud/ai/graph/serializer/agent/JSONStateSerializer.java
@@ -1,4 +1,4 @@
-package com.alibaba.cloud.ai.graph.practice.insurance_sale;
+package com.alibaba.cloud.ai.graph.serializer.agent;
 
 import com.alibaba.cloud.ai.graph.serializer.plain_text.PlainTextStateSerializer;
 import com.alibaba.cloud.ai.graph.state.NodeState;
@@ -6,35 +6,30 @@ import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.PropertyAccessor;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
-import dev.ai.alibaba.samples.executor.AgentAction;
-import dev.ai.alibaba.samples.executor.AgentFinish;
-import dev.ai.alibaba.samples.executor.AgentOutcome;
-import dev.ai.alibaba.samples.executor.IntermediateStep;
 import lombok.NonNull;
 
 import java.io.IOException;
 import java.io.ObjectInput;
 import java.io.ObjectOutput;
 
-public class IsJSONStateSerializer extends PlainTextStateSerializer {
+public class JSONStateSerializer extends PlainTextStateSerializer {
 
 	final ObjectMapper objectMapper;
 
-	public IsJSONStateSerializer() {
+	public JSONStateSerializer() {
 		this(new ObjectMapper());
 	}
 
-	public IsJSONStateSerializer(@NonNull ObjectMapper objectMapper) {
+	public JSONStateSerializer(@NonNull ObjectMapper objectMapper) {
 		super(NodeState::new);
 		this.objectMapper = objectMapper;
 		this.objectMapper.setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY);
 
 		var module = new SimpleModule();
-		module.addDeserializer(NodeState.class, new IsStateDeserializer());
-		module.addDeserializer(AgentOutcome.class, new IsAgentOutcomeDeserializer());
-		module.addDeserializer(AgentAction.class, new IsAgentActionDeserializer());
-		module.addDeserializer(AgentFinish.class, new IsAgentFinishDeserializer());
-		module.addDeserializer(IntermediateStep.class, new IsIntermediateStepDeserializer());
+		module.addDeserializer(NodeState.class, new StateDeserializer());
+		module.addDeserializer(AgentOutcome.class, new AgentOutcomeDeserializer());
+		module.addDeserializer(AgentAction.class, new AgentActionDeserializer());
+		module.addDeserializer(AgentFinish.class, new AgentFinishDeserializer());
 
 		objectMapper.registerModule(module);
 	}
