+++ spring-ai-alibaba-graph/spring-ai-alibaba-graph-run/src/test/java/com/alibaba/cloud/ai/graph/practice/insurance_sale/IsAgentService.java
@@ -2,12 +2,8 @@ package com.alibaba.cloud.ai.graph.practice.insurance_sale;
 
 import dev.ai.alibaba.samples.executor.ToolService;
 import org.springframework.ai.chat.client.ChatClient;
-import org.springframework.ai.chat.messages.ToolResponseMessage;
 import org.springframework.ai.chat.model.ChatResponse;
 import org.springframework.ai.model.function.FunctionCallback;
-
-import java.util.List;
-import java.util.Map;
 
 public class IsAgentService {
 
@@ -22,13 +18,6 @@ public class IsAgentService {
 			.defaultFunctions(functions)
 			.build();
 		this.toolService = toolService;
-	}
-
-	private ToolResponseMessage buildToolResponseMessage(IsExecutor.Step intermediateStep) {
-		var toolCall = intermediateStep.action().toolCall();
-		var response = new ToolResponseMessage.ToolResponse(toolCall.id(), toolCall.name(),
-				intermediateStep.observation());
-		return new ToolResponseMessage(List.of(response), Map.of());
 	}
 
 	public ChatResponse execute(String input) {
