--- spring-ai-alibaba-graph/spring-ai-alibaba-graph-samples/src/main/java/dev/ai/alibaba/samples/executor/AgentAction.java
@@ -1,5 +1,0 @@
-package dev.ai.alibaba.samples.executor;
-
-public record AgentAction(String log) {
-
-}
