--- spring-ai-alibaba-graph/spring-ai-alibaba-graph-samples/src/main/java/dev/ai/alibaba/samples/executor/IntermediateStep.java
@@ -1,4 +1,0 @@
-package dev.ai.alibaba.samples.executor;
-
-public record IntermediateStep(AgentAction action, String observation) {
-}
